(maint) Puppet style guide improvements
Before this patch, defined resources do not live in their own manifests
which is a violation of the Puppet style guide.

This patch resolves this issue by moving all defined resources to their
own Puppet manifests.

This patch also resolves issue #5: Bad path for 'ln' command by updating
the exec resource to include `/bin` to the list of paths.